FEAT - add onFocus and onFocusLost to SKInputWithSuggestions
This commit introduces new callbacks, `onFocus` and `onFocusLost`, to the `SKInputWithSuggestions` class. These callbacks are triggered when the input field gains or loses focus, respectively.

Additionally, an `onFocusChange` callback is added to the `SKInputWithSuggestionsVC` interface and implemented in the view proxy. This callback is triggered whenever the focus state of the input field changes.

The `SKInputWithSuggestions` class is now open, allowing for subclassing and customization.

Finally, the Skot library version is bumped to 1.2.43-ua.