WIP - Correct system bar color and add keepScreenOn
This commit fixes an issue where the navigation bar color was not correctly reflecting the system theme (light/dark) when `enableEdgeToEdge` was active. It introduces a separate `navSystemBarStyle` to handle this.

Additionally, a new `keepScreenOn` property is added to `SKScreenView`. When set to true, it prevents the screen from turning off.

The Skot library version is bumped to 1.3.11-ua, and other dependencies are updated:
- kotlin to 2.2.0
- kotlinx-serialization to 1.9.0
- kotlinpoet to 2.2.0
- android_app_compat to 1.7.1
- ktlint to 1.6.0
- gradle to 8.9.3
- sonarqube to 6.2.0.5505